Updated sprites, organized code
Organized some of the code to help make it a bit more readable, as well as adding some table so we don't have to search for values as much. Also updated some of the sprites, mainly the cat and room. Also added a "magical sign" in which it turns green upon an object being pushed over it (it's the basic puzzle for the prototype).