Slightly puzzle tweaks and re-configs
Made the intro puzzle room less cluttered and slightly changed the puzzle. Also added a reset button for the ball puzzle and any future ones (though still needs a function preferably). Added and moved chair props to each room as clues for the button puzzle. Planning to add a note/textbox to give a hint to player