Playtesting Version 2
Implemented the code for the player to go from room-to-room and level-to-level, relies on the "stage_check" function I made a long time ago. Slightly edited the map to make the room transitions easier as well as to change the visuals slightly. Moved the "level" variable into the global tab.