Flower Puzzle, Updated Forest, and "debug_update"
Added in the functions related to the flower puzzle in the forest level, adjusting the required classes and sprites. Changed the flowers in the forest to be a hint. "debug_update" is added so we don't have to manually change stuff in init, such as cat collision. Also makes sure we don't deploy the game with a the wrong values